Add files via upload
This is the wind information for the scenarios that you were keen to model.  It took a while longer than expected to confirm the information. You mentioned that you wanted to model some scenarios with high and low wind resources from different years. As you can see from the wind capacity factor (version 1).xlsx, I have summarised to the results from the three lowest wind energy to the three highest wind energy years. There appears to be no link between the regional hurricanes to the wind resource in Barbados. For one Barbados is not usually struck by hurricanes due to our geography the hurricanes are formed but they do not impact us directly.  

What can be seen is the during the hurricane season, you could make plots to make show how high or low the wind resource and maybe compare the energy production during different years ranging between the good and bad years for the wind resource. 
 
So 2002 is the best year with 
2010 and 2014 being the two worst years with 2010 being the year that the island was slightly impacted by a hurricane. tell me, what you think. 

the data have all be normalised already be dividing by the maximum value of the energy production of one hour of the year